+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mfettal <mfettal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/16 15:54:45 by mfettal           #+#    #+#             */
-/*   Updated: 2023/04/17 22:02:51 by mfettal          ###   ########.fr       */
+/*   Updated: 2023/04/23 19:23:19 by mfettal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,11 +19,11 @@
 # include <unistd.h>
 
 int		ft_printf(const char *s, ...);
-void	ft_putchar(int c, int *len);
-void	ft_putstr(char *s, int *len);
-void	ft_putnbr(int n, int *len);
-void	ft_putunbr(unsigned int n, int *len);
-void	ft_hexanbr(unsigned int nbr, char n, int *len);
-void	ft_pointer(int *adr, int *len);
+void	ft_putchar(int c, long *len);
+void	ft_putstr(char *s, long *len);
+void	ft_putnbr(int n, long *len);
+void	ft_putunbr(unsigned int n, long *len);
+void	ft_hexanbr(unsigned int nbr, char n, long *len);
+void	ft_pointer(unsigned long *adr, long *len);
 
 #endif
